Move loaders namespace to datasets for API consistency (#711)
* Move loaders namespace to datasets for API consistency

* Reformat

* make commit
