+++ keras_cv/__init__.py
@@ -18,8 +18,8 @@ from keras_cv import version_check
 version_check.check_tf_version()
 # isort:on
 
+from keras_cv import datasets
 from keras_cv import layers
-from keras_cv import loaders
 from keras_cv import losses
 from keras_cv import metrics
 from keras_cv import models
