+++ keras_cv/datasets/pascal_voc/__init__.py
@@ -11,4 +11,4 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from keras_cv.loaders.pascal_voc.load import load
+from keras_cv.datasets.pascal_voc.load import load
